Rename AppVisor activity and move createShortcut to AppVisor
