+++ addon/analytiks-appvisor/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:theme="@style/Theme.AnalyticsApp">
         <activity
-            android:name="com.analytiks.addon.appvisor.ui.MainActivity"
+            android:name="com.analytiks.addon.appvisor.ui.AppVisorActivity"
             android:label="@string/app_name"
             android:launchMode="singleTask"
             android:exported="true"
