+++ addon/analytiks-appvisor/src/main/java/com/analytiks/addon/appvisor/ui/AppVisorActivity.kt
@@ -1,27 +1,31 @@
 package com.analytiks.addon.appvisor.ui
 
-import android.content.Context
 import android.content.Intent
-import android.content.pm.ShortcutInfo
-import android.content.pm.ShortcutManager
-import android.graphics.drawable.Icon
 import android.net.Uri
-import android.os.Build
 import android.os.Bundle
-import android.util.Log
+import android.view.View
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.getSystemService
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.analytiks.addon.appvisor.R
 import com.analytiks.addon.appvisor.databinding.ActivityVisorBinding
 
 
-class MainActivity : AppCompatActivity() {
+class AppVisorActivity : AppCompatActivity() {
     private lateinit var binding: ActivityVisorBinding
     private val adapter by lazy {
         EventsAdapter {}
+    }
+
+    private fun updateUiState(emptyState: Boolean) {
+        if (emptyState) {
+            binding.visorsRecyclerView.visibility = View.GONE
+            binding.emptyStateLayout.visibility = View.VISIBLE
+        } else {
+            binding.visorsRecyclerView.visibility = View.VISIBLE
+            binding.emptyStateLayout.visibility = View.GONE
+        }
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -34,6 +38,7 @@ class MainActivity : AppCompatActivity() {
         binding.visorsRecyclerView.adapter = adapter
 
         AppVisorDataCollector.getInstance().getEvents().observe(this) { logs ->
+            updateUiState(logs.isEmpty())
             adapter.submitList(logs)
         }
 
@@ -53,41 +58,8 @@ class MainActivity : AppCompatActivity() {
     }
 
     companion object {
-
         fun initialize(): LoggingAnalytiksInterceptor {
             return LoggingAnalytiksInterceptor(collector = AppVisorDataCollector.getInstance())
-        }
-
-        fun createShortcut(context: Context) {
-            if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N_MR1) {
-                return
-            }
-
-            val shortcutManager = context.getSystemService<ShortcutManager>() ?: return
-            if (shortcutManager.dynamicShortcuts.any { it.id == "APP_VISOR_ID" }) {
-                return
-            }
-
-            val shortcut =
-                ShortcutInfo.Builder(context, "APP_VISOR_ID")
-                    .setShortLabel("Analytiks Visor")
-                    .setLongLabel("Analytiks App visor - Analyze your app events")
-                    .setIcon(Icon.createWithResource(context, R.mipmap.ic_launcher))
-                    .setIntent(getLaunchIntent(context).setAction(Intent.ACTION_VIEW))
-                    .build()
-            try {
-                shortcutManager.addDynamicShortcuts(listOf(shortcut))
-            } catch (e: IllegalArgumentException) {
-                Log.d("ShortcutManagerTag", "ShortcutManager addDynamicShortcuts failed $e")
-            } catch (e: IllegalStateException) {
-                Log.d("ShortcutManagerTag", "ShortcutManager addDynamicShortcuts failed $e")
-            }
-        }
-
-        fun getLaunchIntent(context: Context): Intent {
-            return Intent(context, MainActivity::class.java)
-                .setFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-
         }
     }
 }
